requirements.txt
django-webpack-loader==0.6.0
pytz==2018.7